arm64: dts: rockchip: Enable USB3 for rk3328 Rock64
Git-commit: 88f0bc830c52f7c6015ea5b174997466ea208c3e
References: bsn#19
Patch-mainline: v5.10.51

commit bbac8bd65f5402281cb7b0452c1c5f367387b459 upstream.

Enable USB3 nodes for the rk3328-based PINE Rock64 board.

The separate power regulator is not added as it is controlled by the
same GPIO line as the existing VBUS regulators, so it is already
enabled. Also there is no port representation to tie the regulator to.

[wens@csie.org: Rebased onto v5.12]

Signed-off-by: Cameron Nemo <cnemo@tutanota.com>
[wens@csie.org: Rewrote commit message]
Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Link: https://lore.kernel.org/r/20210504083616.9654-2-wens@kernel.org
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.51-0137-arm64-dts-rockchip-Enable-USB3-for-rk3328-Rock64.patch