platform/x86: touchscreen_dmi: Add an extra entry for the
Git-commit: d4801889d61f3e9c41dd82a90ecdc14b6854393d
 upside down Goodix touchscreen on Teclast X89 tablets
References: bsn#19
Patch-mainline: v5.10.50

[ Upstream commit a22e3803f2a4d947ff0083a9448a169269ea0f62 ]

Teclast X89 tablets come in 2 versions, with Windows pre-installed and with
Android pre-installed. These 2 versions have different DMI strings.

Add a match for the DMI strings used by the Android version BIOS.

Note the Android version BIOS has a bug in the DSDT where no IRQ is
provided, so for the touchscreen to work a DSDT override fixing this
is necessary as well.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://lore.kernel.org/r/20210504185746.175461-4-hdegoede@redhat.com
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.50-0174-platform-x86-touchscreen_dmi-Add-an-extra-entry-for-.patch